requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap3==14.2.0
django-braces==1.14.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-model-utils==4.1.1
django-registration-redux==2.9
django-rest-framework==0.1.0
django-star-ratings==0.9.1
djangorestframework==3.9.4
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
pyuploadcare==2.7.0
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
urllib3==1.26.2
whitenoise==5.2.0
zipp==3.4.0
